back_python/requirements.txt
# Web框架
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6

# 数据处理
pyspark==3.5.0
kafka-python==2.0.2

# 计算机视觉
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# 深度学习（可选，用于YOLO模型）
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196  # YOLOv8

# 数据库连接
pymysql==1.1.0
sqlalchemy==2.0.23

# 工具类
python-dotenv==1.0.0
# eventlet==0.33.3  # 已移除，改用threading模式

